1.4.3: add missing enumTags support to tiles

+++ Engine/AM2E/Levels/LDtk/LDtkTileInstance.cs
@@ -29,5 +29,5 @@ public struct LDtkTileInstance
     /// The *Tile ID* in the corresponding tileset.
     /// </summary>
     [JsonProperty("t")]
-    public long T { get; set; }
+    public int T { get; set; }
 }
